+++ validation-of-forms/lib/assets/colors.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+
+/// Цвета приложения
+
+class AppColors {
+  static const colorBly = Color(0xFF414657);
+  static const colorGrayText = Color(0xFFAFB2BC);
+  static const colorwhite = Color(0xFFFFFFFF);
+  static const colorPink = Color(0xFFff8087);
+  static const colorGray = Color(0xFFeaeceb);
+  static const colorLigthBly = Color(0xFFd3d5d8);
+  static const colorLightRed = Color(0xFFfd4c56);
+}
