+++ validation-of-forms/lib/assets/strings.dart
@@ -1,0 +1,14 @@
+/// Строки приложения
+
+class AppStrings {
+  static const petName = 'Имя питомца';
+  static const petBirsday = 'День рождения питомца';
+  static const petWeight = 'Вес, кг';
+  static const email = 'Почта хозяина';
+  static const vaccinations = 'Сделаны прививки от:';
+  static const rabies = 'бешенства';
+  static const covid = 'ковида';
+  static const malaria = 'малярии';
+  static const dateVaccinations = 'Дата последней прививки';
+  static const buttonSend = 'Отправить';
+}
