+++ validation-of-forms/lib/domain/models/validators.dart
@@ -1,5 +1,3 @@
-
-
 String? nameValidate(value) {
   if ((value.length < 3 || value.length > 20)) {
     return 'Укажите имя питомца от 3 до 20 символов';
@@ -8,8 +6,7 @@ String? nameValidate(value) {
 }
 
 String? weightValidate(value) {
-  if (value.isEmpty ||
-      int.parse(value) < 1) {
+  if (value.isEmpty ||  int.parse(value) < 1) {
     return 'Укажите вес, больше 0 кг';
   }
   return null;
@@ -25,5 +22,7 @@ String? dateValidate(value) {
 String? emailValidate(value) {
   const pattern = r"^[a-zA-Z0-9.]+@[a-zA-Z0-9]+\.[a-zA-Z]+";
   final regex = RegExp(pattern);
-  return value!.isNotEmpty && !regex.hasMatch(value) ? 'Введите email в формате name@mail.com' : null;
+  return value.isNotEmpty && !regex.hasMatch(value)
+      ? 'Введите email в формате name@mail.com'
+      : null;
 }
