+++ validation-of-forms/lib/main.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:surf_flutter_courses_template/assets/styles.dart';
+import 'package:surf_flutter_courses_template/assets/colors.dart';
 import 'package:surf_flutter_courses_template/ui/home_page.dart';
 
 void main() {
