+++ validation-of-forms/lib/ui/model/input_field.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:surf_flutter_courses_template/assets/colors.dart';
 import 'package:surf_flutter_courses_template/assets/styles.dart';
 
 /// Модель поля ввода
