+++ validation-of-forms/lib/ui/model/pet_card_model.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
+import 'package:surf_flutter_courses_template/assets/colors.dart';
 import 'package:surf_flutter_courses_template/assets/styles.dart';
 
 /// Модель карточки питомца
