+++ validation-of-forms/lib/ui/widgets/sent_button_widget.dart
@@ -1,8 +1,11 @@
 import 'package:flutter/material.dart';
+import 'package:surf_flutter_courses_template/assets/colors.dart';
+import 'package:surf_flutter_courses_template/assets/strings.dart';
 import 'package:surf_flutter_courses_template/assets/styles.dart';
+
 /// виджет кнопки отправки данных
 class SentButtonWidget extends StatelessWidget {
-  final void Function() onPressed;
+  final VoidCallback onPressed;
   final bool loading;
   const SentButtonWidget({
     super.key,
